Feat: Implement session management for employee authentication

+++ backend/Features/Auth/AuthController.cs
@@ -160,7 +160,7 @@ namespace backend.Features.Auth
       public int Id { get; set; }
       public string IdentityNumber { get; set; } = string.Empty;
       public string Phone { get; set; } = string.Empty;
-      public float Salary { get; set; }
+      public decimal Salary { get; set; }
       public string Name { get; set; } = string.Empty;
       public string Email { get; set; } = string.Empty;
       public string Role { get; set; } = string.Empty;
